Added an in-memory subscribe/unsubscribe mechanism
The following commit adds a simple subscribe/unsubscribe mechanism
to the bot. In order to do this, we are using the job_queue
feature from the python-telegram-bot.

Also, we had to use memoization on the remote webdriver as
creating multiple browsers slows down and eventually
crashes the application.

Subscribe and unsubscribe is still managed in-memory,
so it is needed for someone to go and implement a database.