Stop reusing the webdriver connection to avoid timeouts
Before, we were re-using the webdriver connection. The problem
is that if the connection becomes idle, the webdriver becomes
unavailable to keep issuing requests.

This patch fixes that by removing memoization. However,
the main problem of having multiple connections dealing with
the webdriver needs to be addressed. A TODO was placed to do
so.