Add more information about appointments available by date
We have shown that the bot can detect whether or not there
are appointments available. However, how many appointments
are available in which date?

The following patch adds the ability to extract extra
appointment from the webpage. The information is encapsulated
in a new dataclass. Currently, only the date
and the number of appointments in such date is being capture.
Of course, we can expand and include the timeslots.