Remove .env file from being tracked.
In order to keep the secrets from the environment,
I am removed .env from being tracked. This will allow
us to keep the secrets in a file instead
of always having to define an environment variable.

We are also adding .env to the .gitignore file.

Documentation regarding how to start the container
was also updated.